lab_6/db_server/university/views.py
from rest_framework.viewsets import ModelViewSet

from .serializers import StudentSerializer, GroupSerializer, TeacherSerializer, FacultySerializer
from .models import Student, Group, Teacher, Faculty


class StudentViewSet(ModelViewSet):
    queryset = Student.objects.all()
    serializer_class = StudentSerializer

class GroupViewSet(ModelViewSet):
    queryset = Group.objects.all()
    serializer_class = GroupSerializer

class TeacherViewSet(ModelViewSet):
    queryset = Teacher.objects.all()
    serializer_class = TeacherSerializer

class FacultyViewSet(ModelViewSet):
    queryset = Faculty.objects.all()
    serializer_class = FacultySerializer