lab_6/db_server/requirements.txt
Django==3.0.7
djangorestframework
django-cors-headers